Add move site ifo alert
