+++ pages/[areaId]/[chapterId].tsx
@@ -1,17 +1,14 @@
-import {Clear, Map, Search, Tag} from "@mui/icons-material";
-import {Box, Button, Container, IconButton, Paper, Tab, Tabs, TextField, Typography} from "@mui/material";
-import Link from "next/link";
+import {Container} from "@mui/material";
 import {useRouter} from "next/router";
 import {GetStaticPaths, GetStaticProps} from "next/types";
 import {CampPage} from "pages/_app";
 import {ParsedUrlQuery} from "querystring";
-import {createElement, Fragment, useEffect, useState} from "react";
-import {ChapterGridView, ChapterListView, filterCheckpoints} from "~/modules/chapter";
+import {Fragment, useEffect, useState} from "react";
+import {filterCheckpoints} from "~/modules/chapter";
 import {ChapterBreadcrumbs} from "~/modules/chapter/Breadcrumbs";
 import {ChapterHeaderImage} from "~/modules/chapter/HeaderImage";
 import {HeaderNav} from "~/modules/chapter/HeaderNav";
 import {AreaData, ChapterData, ChapterNav, CheckpointData, SideData} from "~/modules/chapter/types";
-import {pluralize} from "~/modules/common/pluralize";
 import {VALID_AREAS} from "~/modules/data/validAreas";
 import {fetchArea, getChapterImageUrl} from "~/modules/fetch/dataApi";
 import {CampHead} from "~/modules/head/CampHead";
@@ -69,89 +66,9 @@ const ChapterPage: CampPage<ChapterProps> = ({area, chapter, sides, prevChapter,
           }}
         />
         <ChapterHeaderImage area={area} chapter={chapter} />
-        {side && (
-          <Paper elevation={1} sx={{display: "flex", alignItems: "center", justifyContent: "space-between", p: 1, mt: 2, mb: 2}}>
-          <Typography component="div" variant="body1" color="text.secondary" textAlign="center">
-              {pluralize(side.roomCount, "room")}
-            </Typography>
-            <Link passHref href={`/map/${area.id}/${chapter.id}/${side.id}`}>
-              <Button variant="contained" endIcon={<Map/>}>Level Map</Button>
-            </Link>
-          </Paper>
-        )}
-        <TextField
-          fullWidth
-          placeholder="Search rooms"
-          autoComplete="off"
-          variant="standard"
-          value={searchValue}
-          InputProps={{
-            endAdornment: (
-              <Box display="flex" alignItems="center" gap={0.5} margin={0.5}>
-                <IconButton
-                  size="small"
-                  onClick={() => setSearchValue("")}
-                  aria-label="clear search"
-                >
-                  <Clear />
-                </IconButton>
-                <Search color="primary" />
-              </Box>
-            ),
-          }}
-          onChange={event => setSearchValue(event.target.value)}
-          onKeyDown={event => {
-            if (event.key === "Enter" && side !== undefined) {
-              setCheckpoints(filterCheckpoints(searchValue, side, true))
-            }
-          }}
-          aria-label="search rooms"
-          sx={{marginTop: 2, marginBottom: 2}}
-        />
-        <Tabs variant="fullWidth" value={side} onChange={(_, newSide) => setSide(newSide)} sx={{mb: 2}}>
-          {sides.map(side => <Tab key={side.name} value={side} label={`${side.name}-side`}/>)}
-        </Tabs>
-        {side && (
-          <Fragment>
-            {Boolean(searchValue) && checkpoints && checkpoints.length === 0 && (
-              <Box display="flex" justifyContent="center" padding={2}>
-                <Typography component="div" fontSize="large" color="text.secondary">{`No rooms found for '${searchValue}'`}</Typography>
-              </Box>
-            )}
-            {checkpoints.map(checkpoint => (
-              <Box key={checkpoint.name} sx={{display: "flex", flexDirection: "column", marginBottom: 2, padding: 0}}>
-                <Typography
-                  id={checkpoint.name}
-                  component="a"
-                  href={`#${checkpoint.name}`}
-                  variant="h5"
-                  color="text.secondary"
-                  mt={1}
-                  mb={1}
-                  display="flex"
-                  alignItems="center"
-                  sx={{
-                    ":hover": {
-                      textDecoration: "underline",
-                      "#anchor-link": {
-                        display: "block",
-                      }
-                    }
-                  }}
-                >
-                  {checkpoint.name}
-                  <Tag id="anchor-link" fontSize="small" sx={{display: "none", ml: 0.5}} />
-                </Typography>
-                {createElement(settings.listMode ? ChapterListView : ChapterGridView, {
-                  areaId: area.id,
-                  chapterId: chapter.id,
-                  sideId: side.id,
-                  checkpoint,
-                })}
-              </Box>
-            ))}
-          </Fragment>
-        )}
+        {sides.map(({id, name, checkpoints, roomCount}) => (
+          <></>
+        ))}
       </Container>
     </Fragment>
   );
