+++ pages/index.tsx
@@ -1,5 +1,6 @@
-import {Avatar, Box, Chip, Container, Paper, Typography} from "@mui/material";
+import {Alert, Avatar, Box, Chip, Container, Paper, Typography} from "@mui/material";
 import {GetStaticProps} from "next";
+import Link from "next/link";
 import {Fragment} from "react";
 import {Area} from "~/modules/data/dataTypes";
 import {fetchArea, getRootImageUrl} from "~/modules/fetch/dataApi";
@@ -45,6 +46,19 @@ export const HomePage: CampPage<AreaProps> = ({area, chapters}) => {
             />
           </Box>
         </Paper>
+        <Alert
+          severity="info"
+          variant="outlined"
+          sx={{mt: 2}}
+        >
+          {"Berry Camp has moved to "}
+          <strong>
+            <Link href="https://berrycamp.github.io">
+              https://berrycamp.github.io
+            </Link>
+          </strong>
+          {". Old links will stop working in April so update your bookmarks!"}
+        </Alert>
         <AreaView area={area} chapters={chapters}/>
       </Container>
     </Fragment>
